% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{assign_step_index}
\alias{assign_step_index}
\title{Assign step indices with proper handling of `>` values
values below min_value are assigned a proportional fraction between 0 and 1}
\usage{
assign_step_index(values, scheme, min_value)
}
\description{
Assign step indices with proper handling of `>` values
values below min_value are assigned a proportional fraction between 0 and 1
}
\examples{
mutate(phen, FLU_MIC_step = assign_step_index(FLU_MIC_val.ineq, dilution_scheme))
}
