% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{calc_lambda_gc}
\alias{calc_lambda_gc}
\title{Calculate genomic inflation factor `lambda['gc']`}
\usage{
calc_lambda_gc(
  x,
  two_sided = TRUE,
  p_tail = c("lower", "upper"),
  include_burden_testing = TRUE,
  use_p_val = "p_lrt"
)
}
\arguments{
\item{x}{dataframe from `read_gemma()`}

\item{two_sided}{P-values calculated as two-sided? `TRUE` or `FALSE`}

\item{p_tail}{Interpret p-values as upper (usual LRT p: P(Χ² >= obs)) or lower (p = F(x_obs)) tail}

\item{include_burden_testing}{Include collapsed rare variants in analysis? (`TRUE` or `FALSE`)}

\item{use_p_val}{Allows use of any p value in gemma dataset. Preferably, `"p_lrt"`.}
}
\description{
Calculate genomic inflation factor `lambda['gc']`
}
