% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{gemma2manhattan}
\alias{gemma2manhattan}
\title{Manhattan plot from gemma dataframe (parsed using \code{\link{read_gemma}}).}
\usage{
gemma2manhattan(
  df,
  use_p_val = "p_lrt",
  label_quants = 0.9999,
  label_percentile = NULL,
  y_trans = NULL,
  show_QQ = FALSE,
  coords = TRUE,
  return_annots = FALSE,
  quant_probs = c(0, 0.9, 0.95, 0.99, 0.999, 0.9999, 1),
  ...
)
}
\arguments{
\item{use_p_val}{Allows use of any p value in dataset.
Options: "p_wald", "p_lrt", or "p_score".}

\item{...}{are extra arguments to geom_text_repel}
}
\description{
Get the manhattan plot sorted, associate model results with snpeff annotations.
XXX Expects gt_annots_ann to be present in environment
}
