% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{parse_ANN}
\alias{parse_ANN}
\title{Parses SnpEff annotations from a string, assuming ANN format.}
\usage{
parse_ANN(
  df,
  ann_col = "ANN",
  too_few = "error",
  filter_alt_allele = FALSE,
  AA_abbr = TRUE,
  remove_GT_AAs = TRUE
)
}
\description{
Parses SnpEff annotations from a string, assuming ANN format.
}
\details{
Two main use-cases: in VCF, and in output of vcf2matrix python script
}
