% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{parse_bcf_stats_compare_1iso}
\alias{parse_bcf_stats_compare_1iso}
\title{Parses output from bcftools stats runs.
Designed around runs of exactly 2 input vcf files.}
\usage{
parse_bcf_stats_compare_1iso(filename, return_DP_distr = FALSE)
}
\description{
Parses output from bcftools stats runs.
Designed around runs of exactly 2 input vcf files.
}
