% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{parse_HGVS.p}
\alias{parse_HGVS.p}
\title{Parse HGVS amino acid code to more familiar format.}
\usage{
parse_HGVS.p(df, col = "HGVS.p", remove_GT_AAs = TRUE)
}
\description{
Parse HGVS amino acid code to more familiar format.
}
\details{
The output formatting is: `<AA_ref><AA_pos><AA_alt>.`
Not all HGVS codes are supported for easy conversion,
this was originally designed for simple substitutions.
}
