% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{phen_logistic2numeric}
\alias{phen_logistic2numeric}
\title{Convert true/false into simplified gemma input;
binary phenotypes (0/1) or forced continuous (1/2)}
\usage{
phen_logistic2numeric(x, binary = TRUE)
}
\arguments{
\item{x}{Named boolean vector of phenotypes}
}
\description{
Convert true/false into simplified gemma input;
binary phenotypes (0/1) or forced continuous (1/2)
}
