% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{read_gemma}
\alias{read_gemma}
\title{Read GEMMA file.}
\usage{
read_gemma(filename, use_p_val = "p_lrt", ...)
}
\arguments{
\item{use_p_val}{Select p-value column on which to calculate expected p-value (for QQ plot)}

\item{...}{Extra arguments for `read_tsv()`}
}
\description{
Read GEMMA file.
}
