% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{read_geneo}
\alias{read_geneo}
\title{Parse output from geneo-typing script}
\usage{
read_geneo(filename, gene_name_regex = "(?<=geneo_auto_).*(?=.tsv)")
}
\arguments{
\item{filename}{Name of file output from variant identification script (`~/data/local/scripts/identify_gene_variants.R`)}

\item{gene_name_regex}{Regular expression string to infer gene name from filename}
}
\description{
Parse output from geneo-typing script
}
