% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{read_rel_matrix}
\alias{read_rel_matrix}
\title{Parse relatedness matrix from GEMMA}
\usage{
read_rel_matrix(file, inds, long = TRUE)
}
\arguments{
\item{inds}{a character vector of isolate names}
}
\description{
Parse relatedness matrix from GEMMA
}
