% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{split_ipr_goterms}
\alias{split_ipr_goterms}
\title{Split GO terms into multiple lines}
\usage{
split_ipr_goterms(x, GO_col = "GO_terms")
}
\arguments{
\item{x}{interproscan tsv parsed with \code{\link{read_iprscan}}.}
}
\value{
The same dataframe but separated by GO terms.
}
\description{
Split GO terms into multiple lines
}
