% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{tidy_snpeff}
\alias{tidy_snpeff}
\title{Parses snpeff annotations from VCF into a dataframe.}
\usage{
tidy_snpeff(vcf, AA_abbr = TRUE, remove_GT_AAs = TRUE)
}
\arguments{
\item{vcf}{vcfR object, as from \link[vcfR]{read.vcfR}}

\item{AA_abbr}{Use abbreviations for amino acids.}

\item{remove_GT_AAs}{Remove extra columns with abbreviated amino acid codes.}
}
\description{
Parses snpeff annotations from VCF into a dataframe.
}
\details{
Optionally, parse HGVS amino acide code to more familiar variant codes.
}
