% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{vcf_to_gt_matr}
\alias{vcf_to_gt_matr}
\title{Prepare genotype matrix using VCF or tidy-extracted data.}
\usage{
vcf_to_gt_matr(
  vcf,
  tidy_filt_info = NULL,
  tidy_filt_gts = NULL,
  max_indel_size = 3,
  max_prop_amb_samples = 0.1,
  plot_prop_amb = FALSE,
  coding_only = TRUE,
  ignore_low = FALSE,
  collapse_freq = 0.01,
  collapse_singletons = FALSE,
  separate_singletons = FALSE
)
}
\value{
Matrix with rownames according to individual IDs
}
\description{
Prepare genotype matrix using VCF or tidy-extracted data.
}
\details{
To reduce memory overhead when bug-testing, use prefiltered info and GT
}
