% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{vcfr_to_gwas_gts}
\alias{vcfr_to_gwas_gts}
\title{Filter GTs to only the locations in INFO}
\usage{
vcfr_to_gwas_gts(
  vcf,
  tidy_filt_info,
  max_prop_amb_samples = 0.2,
  plot_prop_amb = FALSE
)
}
\description{
Filter GTs to only the locations in INFO
}
