% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{vcfr_to_gwas_info}
\alias{vcfr_to_gwas_info}
\title{Filter variants down to target loci for GWAS.}
\usage{
vcfr_to_gwas_info(
  vcf,
  max_indel_size = 3,
  coding_only = TRUE,
  ignore_low = FALSE,
  collapse_freq = 0.01
)
}
\description{
Filter variants down to target loci for GWAS.
}
