% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{calc_lambda_gc}
\alias{calc_lambda_gc}
\title{Calculate genomic inflation factor `lambda['gc']`}
\usage{
calc_lambda_gc(df, include_burden_testing = FALSE, use_p_val = "p_lrt")
}
\arguments{
\item{df}{dataframe from `read_gemma()`}

\item{include_burden_testing}{Include collapsed rare variants in analysis? (`TRUE` or `FALSE`)}

\item{use_p_val}{Allows use of any p value in gemma dataset. Preferably, `"p_lrt"`.}
}
\description{
Calculate genomic inflation factor `lambda['gc']`
}
