% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing_helpers.R
\name{read_geneo}
\alias{read_geneo}
\title{Parse output from geneo-typing script}
\usage{
read_geneo(filename, gene_name_regex = "(?<=geneo_auto_).*(?=.tsv)")

read_geneo(filename, gene_name_regex = "(?<=geneo_auto_).*(?=.tsv)")
}
\arguments{
\item{filename}{Path to file containing output from gene-otyping tool}

\item{gene_name_regex}{Extract gene name for columns based on regular expression of filename}
}
\description{
Parse output from geneo-typing script

Parse output from gene-otyping tool
}
